// front-Regra/src/app/components/home/interfaces.ts
export interface RegraData {
  CAB_ID: number;
  CAB_VERBA: string;
  CAB_DESCR: string;
  solicitacao: string;
  dataInc: string;
  crUser: string;
  qtdFunc: string;
  userAprov: string;
  dataAprov: string;
  CAB_STATUS: string;
  userValid: string;
  dataValid: string;
  userCanc: string;
  dataCanc: string;
}